+++ core/common/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @itwin/core-common
 
-This log was last generated on Fri, 24 Jan 2025 08:03:59 GMT and should not be manually modified.
+This log was last generated on Tue, 18 Feb 2025 17:28:32 GMT and should not be manually modified.
+
+## 4.10.7
+Tue, 18 Feb 2025 17:27:03 GMT
+
+### Updates
+
+- Update ECSql reader api to support no stat and error code.
 
 ## 4.10.6
 Fri, 24 Jan 2025 08:02:40 GMT